+++ src/Services/CustomerManager.php
@@ -6,7 +6,7 @@ use DateTimeImmutable;
 use Exception;
 use Shopware\Core\Checkout\Customer\Event\CustomerLoginEvent;
 use Shopware\Core\Checkout\Customer\SalesChannel\AbstractLogoutRoute;
-use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\Routing\Event\SalesChannelContextResolvedEvent;
 use Shopware\Core\Framework\Routing\SalesChannelRequestContextResolver;
 use Shopware\Core\Framework\Validation\DataBag\RequestDataBag;
@@ -22,7 +22,7 @@ class CustomerManager
 {
     private SalesChannelContextRestorer $contextRestorer;
     private EventDispatcherInterface $dispatcher;
-    private EntityRepositoryInterface $customerRepository;
+    private EntityRepository $customerRepository;
     private SalesChannelRequestContextResolver $contextResolver;
     private AbstractLogoutRoute $logoutRoute;
     private SalesChannelContextPersister $contextPersist;
@@ -30,7 +30,7 @@ class CustomerManager
     public function __construct(
         SalesChannelContextRestorer $contextRestorer,
         EventDispatcherInterface $dispatcher,
-        EntityRepositoryInterface $customerRepository,
+        EntityRepository $customerRepository,
         SalesChannelContextPersister $contextPersist,
         SalesChannelRequestContextResolver $contextResolver,
         AbstractLogoutRoute $logoutRoute
@@ -63,7 +63,7 @@ class CustomerManager
 
     public function loginCustomerById(string $customerId, SalesChannelContext $context): SalesChannelContext
     {
-        $this->handleCustomerContext($customerId, $context);
+        $this->extendCustomerTokenLife($context->getToken(), $context->getSalesChannelId(), $customerId);
         $newContext = $this->contextRestorer->restore($customerId, $context);
         $this->customerRepository->update([
             [
@@ -96,13 +96,14 @@ class CustomerManager
      * Since it will attempt to create a new token & syncing it with the App requires
      * the user to log back in, we try to extend the token life if it expired.
      */
-    private function handleCustomerContext(string $customerId, SalesChannelContext $currentContext): void
+    public function extendCustomerTokenLife(string $token, string $channelId, ?string $customerId = null): void
     {
-        $channelId = $currentContext->getSalesChannel()->getId();
-        $customerPayload = $this->contextPersist->load($currentContext->getToken(), $channelId, $customerId);
+        $customerPayload = $this->contextPersist->load($token, $channelId, $customerId);
 
         if ($customerPayload['expired'] ?? null) {
-            $this->contextPersist->save($customerPayload['token'], ['expired' => false], $channelId, $customerId);
+            $newToken = $customerPayload['token'] ?? $token;
+            $newCustomerId = $customerId ?: $customerPayload['customerId'] ?? null;
+            $this->contextPersist->save($newToken, ['expired' => false], $channelId, $newCustomerId);
         }
     }
 }