+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0
 
 - Added `close` button to success purchase page
+- Removed `back` button from success purchase page
 - Removed footer & cookie drawer via CSS
 
 # 0.1.2